added protctd route

+++ backend/controllers/handleFactory.js
@@ -4,6 +4,8 @@ const APIFeatures = require("../utils/apiFeatures");
 
 exports.createOne = (Model) =>
   catchAsync(async (req, res, next) => {
+    if (!req.body.user) req.body.user = req.user;
+    if (req.body.role) delete req.body.role;
     const newQuestion = await Model.create(req.body);
     res.status(201).json({
       status: "success",
@@ -52,10 +54,21 @@ exports.getAll = (Model) =>
     });
   });
 
-exports.deleteOne = (Model) =>
+exports.deleteOne = (Model, mode) =>
   catchAsync(async (req, res, next) => {
-    const doc = await Model.findByIdAndDelete(req.params.id);
+    //! Check if the currently logged in user has created th document
+    const document = await Model.findById(req.params.id);
+    let id_user;
+    if (mode === "answer") id_user = document.user._id;
+    if (mode === "question") id_user = document.user.find((i) => i._id)._id;
 
+    console.log(id_user, req.user._id);
+    if (!id_user.equals(req.user._id))
+      return next(
+        new AppError("You are not allowed to perform this action", 403)
+      );
+
+    const doc = await Model.findByIdAndDelete(req.params.id);
     if (!doc) {
       return next(new AppError("No document found with that id", 404));
     }
@@ -65,22 +78,29 @@ exports.deleteOne = (Model) =>
     });
   });
 
-exports.updateOne = (Model) =>
+exports.updateOne = (Model, mode) =>
   catchAsync(async (req, res, next) => {
-    const question = await Model.findByIdAndUpdate(
-      req.params.id,
-      { question: req.body.question },
-      {
-        new: true,
-        runValidators: true,
-      }
-    );
+    //! Check if the currently logged in user has created th document
+    const document = await Model.findById(req.params.id);
+    let id_user;
+    if (mode === "answer") id_user = document.user._id;
+    if (mode === "question") id_user = document.user.find((i) => i._id)._id;
 
-    if (!question) {
+    console.log(id_user, req.user._id);
+    if (!id_user.equals(req.user._id))
+      return next(
+        new AppError("You are not allowed to perform this action", 403)
+      );
+    const doc = await Model.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+      runValidators: true,
+    });
+
+    if (!doc) {
       return next(new AppError("No document found with that id", 404));
     }
     res.status(201).json({
       status: "success",
-      data: question,
+      data: doc,
     });
   });
